requirements.txt
# DocFlow - Document Processing Platform
# Python 3.11+ required

# Core dependencies
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0
httpx>=0.25.0

# LLM Integration
mistralai>=0.0.7

# PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0

# Database (PostgreSQL)
asyncpg>=0.29.0

# Async Support
aiofiles>=23.2.0
aiocache>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
python-dotenv>=1.0.0
rich>=13.0.0

# Type hints
typing-extensions>=4.8.0
